removed entity and just reuturned status from job apis

+++ sql-scripts/create-job-specs-table.sql
@@ -1,5 +1,5 @@
 use TheITCrowd_MiaG;
 
-CREATE TABLE JobSpecs (jobSpecId INT AUTO_INCREMENT, jobSpecName VARCHAR(255) NOT NULL, PRIMARY KEY (jobSpecId));
+CREATE TABLE JobSpecs (jobSpecId SMALLINT AUTO_INCREMENT, jobSpecName VARCHAR(255) NOT NULL, PRIMARY KEY (jobSpecId));
 
 INSERT INTO JobSpecs (jobSpecName) VALUES ("Data Engineer");
